+++ ProjectBanHangNew/src/main/java/com/example/Entity/OrderItemDTOShopCart.java
@@ -15,7 +15,7 @@ public class OrderItemDTOShopCart {
 
 	private String product_name;
 
-	private BigDecimal product_price;
+	private Integer product_price;
 
 	private Integer quantity;
